+++ EegLinearFilter/src/main.cpp
@@ -14,7 +14,7 @@ int main(int argc, const char * argv[]) {
     
     try {
         // TODO - padding is missing
-        std::vector<int16_t> allData = loadEdfData(filePath);
+        std::vector<float> allData = loadEdfData(filePath);
         
         std::cout << "Načtená data (prvních 100 vzorků z celkových " << allData.size() << "):" << std::endl;
         for (size_t i = 0; i < 100 && i < allData.size(); ++i) {
